feat: enhance recommendation system with cluster-aware deduplication and source tracking
- Add source field to RecommendationCard and API responses
- Implement cluster-aware recommendation fetching in BookQueryRepository
  - Resolve cluster member IDs for broader recommendations
  - Merge recommendations by source priority (pipeline > same_author > same_category)
- Add two-pass search deduplication (cluster-based + title/author-based)
- Resolve primary editions for recommendation persistence
- Fix integer overflow in cover image dimension calculations (width*height)
- Expand Google Books hostname validation to include googleusercontent.com

Tests updated to reflect new source field and cover suppression metadata.

+++ src/main/java/com/williamcallahan/book_recommendation_engine/service/PostgresBookRepository.java
@@ -430,7 +430,7 @@ public class PostgresBookRepository {
                 ORDER BY
                     -- Prioritize actual quality metrics (Task #5)
                     COALESCE(is_high_resolution, false) DESC,
-                    COALESCE(width * height, 0) DESC,
+                    COALESCE((width::bigint * height::bigint), 0) DESC,
                     -- Fall back to image type as tiebreaker
                     CASE image_type
                         WHEN 'extraLarge' THEN 1
